% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/facet_ragged_rows.R
\name{facet_ragged_rows}
\alias{facet_ragged_rows}
\title{Lay out panels in a grid of ragged rows}
\usage{
facet_ragged_rows(rows, cols, labeller = "label_value")
}
\arguments{
\item{rows, cols}{A set of variables or expressions quoted by \code{\link[ggplot2:vars]{ggplot2::vars()}},
the combinations of which define panels to be inlcuded in the grid. Panels
for \code{cols} are wrapped independently within \code{rows} to form a ragged grid.}

\item{labeller}{A function that takes one data frame of labels and
returns a list or data frame of character vectors. Each input
column corresponds to one factor. Thus there will be more than
one with \code{vars(cyl, am)}. Each output
column gets displayed as one separate line in the strip
label. This function should inherit from the "labeller" S3 class
for compatibility with \code{\link[ggplot2:labeller]{labeller()}}. You can use different labeling
functions for different kind of labels, for example use \code{\link[ggplot2:label_parsed]{label_parsed()}} for
formatting facet labels. \code{\link[ggplot2:label_value]{label_value()}} is used by default,
check it for more details and pointers to other options.}
}
\description{
Lay out panels in a grid of ragged rows
}
\examples{
p <- ggplot(Indometh, aes(time, conc)) + geom_line()

# Ragged grid with cohorts on separate rows
p + facet_ragged_rows(
  vars(Cohort = 1 + Subject \%in\% 3:6),
  vars(Subject = as.character(Subject)),
  labeller = label_both
)
}
