% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggragged-package.R
\docType{package}
\name{ggragged-package}
\alias{ggragged}
\alias{ggragged-package}
\title{ggragged: Ragged Grids for ggplot2}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Faceting functions for defining panel layouts arranged in a grid with ragged edges in ggplot2. These can be useful, for example, to represent nested or partially crossed relationships between faceting variables.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/mikmart/ggragged}
  \item \url{https://mikmart.github.io/ggragged/}
  \item Report bugs at \url{https://github.com/mikmart/ggragged/issues}
}

}
\author{
\strong{Maintainer}: Mikko Marttila \email{mikkmart@pm.me}

}
\keyword{internal}
